Utowrzenie metod asnyc

+++ PortalRandkowy.API/Controllers/ValuesController.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 using PortalRandkowy.API.Data;
 using PortalRandkowy.API.Models;
 
@@ -20,47 +21,49 @@ namespace PortalRandkowy.API.Controllers
         }
         // GET api/values
         [HttpGet]
-        public IActionResult GetValues()
+        public async Task<IActionResult> GetValues()
         {
-            var values = this.context.Values.ToList();
+            var values = await this.context.Values.ToListAsync();
             return Ok(values);
         }
 
         // GET api/values/5
         [HttpGet("{id}")]
-        public IActionResult GetValue(int id)
+        public async Task<IActionResult> GetValue(int id)
         {
-            var value = this.context.Values.FirstOrDefault(x=>x.Id == id);
+            var value = await this.context.Values.FirstOrDefaultAsync(x=>x.Id == id);
             return Ok(value);
         }
 
         // POST api/values
         [HttpPost]
-        public IActionResult AddValue([FromBody] Value value)
+        public async Task<IActionResult> AddValue([FromBody] Value value)
         {
             this.context.Values.Add(value);
-            this.context.SaveChanges();
+            await this.context.SaveChangesAsync();
             return Ok(value);
         }
 
         // PUT api/values/5
         [HttpPut("{id}")]
-        public IActionResult EditValue(int id, [FromBody] Value value)
+        public async Task<IActionResult> EditValue(int id, [FromBody] Value value)
         {
-            var data = this.context.Values.Find(id);
+            var data = await this.context.Values.FindAsync(id);
             data.Name = value.Name;
             this.context.Values.Update(data);
-            this.context.SaveChanges();
+            await this.context.SaveChangesAsync();
             return Ok(data);
         }
 
         // DELETE api/values/5
         [HttpDelete("{id}")]
-        public IActionResult DeleteValue(int id)
+        public async Task<IActionResult> DeleteValue(int id)
         {
-            var data = this.context.Values.Find(id);
+            var data = await this.context.Values.FindAsync(id);
+            if(data == null)
+                return NoContent();
             this.context.Values.Remove(data);
-            this.context.SaveChanges();
+            await this.context.SaveChangesAsync();
             return Ok(data);
         }
     }
